+++ packages/ranger-hooks/src/index.ts
@@ -1,2 +1,3 @@
 export { useCookie } from './cookie'
+export { usePwaInstall } from './pwa'
 export * from 'ahooks'