+++ packages/ranger-ui/src/InstallTrigger/index.ts
@@ -1,0 +1,2 @@
+export type { InstallTriggerProps } from './installTrigger'
+export { default as InstallTrigger } from './installTrigger'