+++ packages/ranger-adobe-edege/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ranger-theme/adobe-edege
 
+## 0.2.0
+
+### Minor Changes
+
+- adobe edege bin command
+
 ## 0.1.0
 
 ### Minor Changes