+++ packages/ranger-hooks/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @ranger-theme/hooks
 
+## 0.7.0
+
+### Minor Changes
+
+- adobe edege bin command
+
+### Patch Changes
+
+- Updated dependencies
+  - @ranger-theme/utils@0.9.0
+
 ## 0.6.0
 
 ### Minor Changes