+++ packages/ranger-tracking/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ranger-theme/tracking
 
+## 0.4.0
+
+### Minor Changes
+
+- adobe edege bin command
+
 ## 0.3.0
 
 ### Minor Changes